requirements.txt
# Function dependencies, for example:
# package>=version
APScheduler==3.9.1.post1
autopep8==2.0.1
beautifulsoup4==4.11.1
bottle==0.12.23
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google==3.0.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
greenlet==2.0.1
gspread==5.7.2
gspread-dataframe==3.3.0
idna==3.4
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
tomli==2.0.1
tweepy==4.12.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
openai==0.27.4